Return request.csrf_tag as a Markup object
This indicates that it shouldn't be escaped so that it's not necessary
to explicitly indicate this to templating engines.

E.g., in Mako, it's now possible to say `${request.csrf_tag}` instead of
`${request.csrf_tag | n}`.

I'm not sure if this works with other templating engines, but it
shouldn't cause issues where it's not supported (since Markup is just
a subclass of str).