Improve URL var setup
There is one big, back imcompatible change: URL vars are now delimited
by `<...>` instead of `{...}`. This is because angle brackets are
unlikely to occur in the regular expressions for URL vars while curly
brackets are quite likely to. The upshot is that we can simplify the
URL var regex quite a bit by using `<...>`.

In addition, a converter can now be specified like so:

    /<(int)id:\d+>

When a URL matches a mounted resource, URL vars will be converted
automatically, and when generating URLs for a mounted resource, the vars
passed in will be validated by ensuring they can be converted (although
perhaps this is too restrictive?).

Started adding tests for MountedResource, especially relating to these
changes.