Include representations in 3xx responses when appropriate
If a resource method returns data for a 3xx response, that data should
be returned to the client. Previously, all such data was discarded.

In particular, responses with a 300 status code "SHOULD include an
entity containing a list of resource characteristics and location(s)
from which the user or user agent can choose the one most appropriate."

The spec also recommends that most other 3xx responses SHOULD include
"a short hypertext note with a hyperlink to the new URI(s)."