Evaluate Application._handlers property when app is created
This will cause the modules where the handlers are defined to be
imported, which will expose import-time errors early (instead of waiting
for the first request).

A related idea is to issue a request for / after the app is created, but
that's not as straightforward. It's not clear how to do that in a way
that's application/settings-agnostic.