Add CORS handler
This is mainly intended for use in development with the dev server. When
the dev server is started with --enable-permissive-cors, CORS will be
allowed from any origin for any request method and content type.

This makes it easy for a front end dev server to send AJAX requests
without needing to mess around with setting up a proxy.

TODO: Flesh out a bit to allow all requests from safe origins. E.g., for
      example.com => api.example.com.