Improve handling of exception responses
This builds on the previous commit by ensuring that *all* exception
responses--including 300-level responses--will be handled the same way
(in particular wrt. to CORS).

The `error_handler` function that was added previously was renamed to
`get_exc_response` since it's not really a handler and the new name is
more descriptive.

Amends 7cd6233e1f6139c90b49b580704c4de99d1a380b